main: Generate server certificates
As part of the "no-stdin" design of estampa, this action will be
executed automatically if the server can't find the defined files for
the chain. A conditional has also been added if only one of the files
can be found but not the other. (ex. the certificate is present, but
the key is missing)