config: Simplify settings
Parameters for mailboxes, trust directory, and client certificates have
been unified into a "store" parameter. The structure for data storage
will be automatically handled instead of being explicitly set by the
admin. In the case that the admin needs to create a server with
mailboxes for different users in a single machine (ex. pubnixes), they
can set the ownership to each user, and keep write access enabled for
Estampa using a group.