meta: Bump version to v0.1
Initial release! This version of estampa covers the bare minimums to be
a Misfin implementation. In other words, it only works as a Misfin(B)
mailbox, and can only store messages as Gemtext files. Misfin(C), GMAP,
and the gembox format are yet to be implemented in future releases, but
for now users can read their messages directly from the `mbox/`
directory.