protocol: Implement identifier
This splits off Misfin(B) parsing code into the `protocol` module, and
to keep functionality it adds an "identifier", which will attempt to
parse a request into a type corresponding to each implemented protocol.
For now, that is just Misfin(B), but Misfin(C) and GMAP are in the
roadmap.