mailbox: Message IDs
Messages are now stored with an ID as the file name. This ID is a BLAKE3
hash of the message timestamp, and the sender address, concatenated next
to each other.

The extension for message files has also been changed to .msfn from
.gmi.

Lastly, as this is a big change in terms of format changes, a new
versioning system has been added to the "store" folder, which will later
be complemented with a migration module. This consists of a file
contained at the root of the store, named ".version", which will store
a number denoting the version (ex, "2").