mailbox: Restructure mailbox module
Interactions with mailboxes are now handled through a new Mailbox
struct, instead of the previously used Message struct. The mailbox is
issued by the Config struct, using the mailbox() method. This allows
for mailbox calls to be a lot more readable, and follow a more object
oriented approach. Also, this adds some methods under the same struct
for a future GMAP implementation. New errors still need to be added,
and the handler requires urgent cleanup.