handler: Map errors instead of pattern matching
Boiler(plate)? Who 'er?

Jokes aside, this removes a lot of the match blocks and instead uses the
map_or_else() method in Rust to handle errors without adding boilerplate
code. In other words, get results with Try as one would in a function
that returns a Result, then map errors if any are present at the very
end of the handler.